Move drug fetching to CLI.
Fixes #47
