+++ cli/Cargo.toml
@@ -9,13 +9,11 @@ name = "extract-drugs"
 path = "src/main.rs"
 
 [dependencies]
-csv = "1.1.6"
 clap = { version = "3.1.9", features = ["derive", "cargo"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+csv = "1.1.6"
 indicatif = "0.16.2"
-walkdir = "2.3.2"
 dialoguer = "0.10.0"
-console = "0.15.0"
+reqwest = {version = "0.11.10", features = ["json", "blocking"]}
 drug-extraction-core = { path = "../core" }
-futures = "0.3.21"
